# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'NewsNet API'
  description: 'Web Resources Specification (A7) for NewsNet'
servers:
  - url: http://lbaw.fe.up.pt
    description: Production server
...
paths:
...
  /login:
    get:
      operationId: R103
      summary: 'R103: Login Form'
      description: 'Provide login form. Access: PUB'
      tags:
        - 'M01: Authentication'
      responses:
        '200':
          description: 'Ok. Show Log-in UI'
    post:
      operationId: R104
      summary: 'R104: User login'
      description: 'Authenticates a user. Access: PUB'
      tags:
        - 'M01: Authentication'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful authentication. Redirect to user profile.'
                  value: '/users/{id}'
                302Error:
                  description: 'Failed authentication. Redirect to login form.'
                  value: '/login'
...
  /api/posts:
    post:
      operationId: R301
      summary: 'R301: Create a New Post'
      description: 'Create a new news post. Access: USR'
      tags:
        - 'M03: News Posts'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
              required:
                - title
                - body
      responses:
        '302':
          description: 'Redirect after creating new post.'
          headers:
            Location:
              schema:
                type: string
              example:
                302Success:
                  description: 'Successful. Post created. Redirect to post.'
                  value: '/posts/{post_id}'
                302Failure:
                  description: 'Failed. Redirect back.'
                  value: '/posts'
...